refactor: changes timer widget directory location
